move http-client files to src/
