+++ packages/http-client/src/auth.ts
