+++ packages/http-client/src/interfaces.ts
